requirements_v3api.txt
git+https://github.com/rancher/client-python.git@master#egg=client-python
boto3==1.10.6
botocore==1.13.6
cryptography==2.8
flake8==3.5.0
invoke==1.2.0
Jinja2==2.10.3
kubernetes==7.0.0
netaddr==0.7.19
paramiko==2.6.0
PyJWT==1.6.4
pytest-repeat==0.7.0
pytest-xdist==1.23.0
pytest==3.8.1
pytest-html==1.20.0
pytest-ordering==0.6
python-digitalocean==1.13.2
python-terraform==0.10.1
PyYAML==5.2
requests==2.22.0
rsa==4.0
websocket-client==0.57.0